Update patch set 1
Patch Set 1: Looks good to me, approved

seriously, add a script that removes all the unneeded files on import, especially build files, but also test files etc if we aren't using them. that took forever to go through because of gerrit slowness and almost all of it was noise.

Patch-set: 1
Label: Code-Review=+2