Update patch set 4
Patch Set 4:

enh@: We're trying to enable the use of libexpat for an sphal for treble (mesa3d, the gpu driver). This is an alternative to blessing the library; it will be emitted shared to /vendor, but it will not be shared with the /system copy.

Patch-set: 4
Reviewer: Gerrit User 1003224 <1003224@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1293738 <1293738@85c56323-6fa9-3386-8a01-6480fb634889>