Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> So, it looks like the previous CL (https://android-review.googlesource.com/c/platform/external/expat/+/720957) was closed without answering to the question that I asked. "Where is the XML file that the gfx HAL is reading?"

This appears to be the driconf XML, which is documented here: https://dri.freedesktop.org/wiki/ConfigurationInfrastructure/

In practice, I doubt this file will ever be opened, it isn't installed by default and we're unlikely to ever need it.

Maybe the best fix of all would be to make the libexpat dependency of mesa optional.

Patch-set: 4
Reviewer: Gerrit User 1293738 <1293738@85c56323-6fa9-3386-8a01-6480fb634889>