Update patch set 4
Patch Set 4:

> My apologies, I thought your queries were only related to that particular change, so when we tried this alternative, I didn't realise the query was still important.
> 
No worries :)

> About the confidence about dual-loading immunity, is a board booting to Android UI with mesa3d good enough example, or would you think there's another way to prove it? I can also try that path, which will be much less invasive on mesa3d.

That's good but not enough. The problem caused by dual loading might happen in some rare cases. For example, we had the similar issue with libutils (which is dual-loaded); the problem happened only when the trace was on. Basically, if the lib (libexpat) is written with an assumption that it is the sole instance that can hold a specific process-wide resource (e.g. a FD), then this is no-go. However, since I am not the author of the lib, I am not sure. So, we are just conservative when deciding whether a lib can be in VNDK-SP or not.

Patch-set: 4
Reviewer: Gerrit User 1132673 <1132673@85c56323-6fa9-3386-8a01-6480fb634889>