Update patch set 4
Patch Set 4:

> So I looked into the mesa3d code, and the expat lib seems to be used in 3 additional places:
> - mesa3d utils, and mesa's drivers for broadcom and intel.
> Right now, for Android, all of these are getting built, so I suspect we might not be able to drop the dependency. Admittedly, I don't know how hard is this dependency, so I will write to the authors of all 3 users inside mesa.
> I also haven't got a chance to reach out to expat developers to check about dual-loading, so I will do that as well, and update when I hear back from them.

Ok thanks. I will wait.

In the meantime, I would like to mention that in Android, starting from O-MR1, GL drivers (more generally, a class of HALs that we call Same-Process HALs which include GL drivers) are NOT allowed to access a config file in the vendor partition. This is enforced by the selinux policy. The GL drivers are executed inside the processes for Android Apps and such processes don't have access to files in the vendor partition other than those are explicitly exempted. So even mesa3d keeps the dependency to libexpat, I expect it would be no-op at least in Android.

Patch-set: 4
Reviewer: Gerrit User 1132673 <1132673@85c56323-6fa9-3386-8a01-6480fb634889>