Update patch set 4
Patch Set 4:

So, how was the question to upstream phrased? I think we've moved on from the static vs dynamic discussion. The important point was that the expat code in Mesa for Android does not actually do anything on newer versions because default sepolicy prohibits access from these sphal libraries to 'etc' configs, which is what this dependency was added to read. In general, I'm quite against just hacking Mesa to make it build. If Android cannot/does not use the XML parser, it shouldn't be mandatory to have it in the Mesa build.

Patch-set: 4
Reviewer: Gerrit User 1293738 <1293738@85c56323-6fa9-3386-8a01-6480fb634889>