Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> Hi Sumit, it's been a long time since we heard about these changes, is somebody still working on it?

Apologies - I had to be out due to a family tragedy, and have only returned last week. I will restart work on this.

I was hoping to be at LPC to discuss the mechanism to do this in an 'upstream-friendly' way, but I couldn't travel. Did you guys get a chance to discuss it there?

Basically, as a recap, according to Jiyong, , the preferences are

1) make mesa3d to not depend on libexpat for Android
2) make mesa3d to statically link to libexpat
3) make libexpat a vendor lib (this CL)
4) make libexpat a VNDK-SP library

I (and Amit P, in my absence) tried to examine how to replace expat usage in mesa, but it looks like there are multiple usages in mesa3d (via mesa shim layer), which seems to provide 'default' values as well. As such, 1 option for implementing 1) would be to provide Android-specific replacement for the mesa3d wrappers - am not sure upstream would be happy about it much.
2 is still a feasible option, and while testing earlier on dragonboard, I realised that there don't seem to be any config files required to be stored or accessed. Thoughts?

Patch-set: 4
Reviewer: Gerrit User 1052193 <1052193@85c56323-6fa9-3386-8a01-6480fb634889>